derived quantities test package
